@@ -4,10 +4,10 @@
   "sideEffects": false,
   "type": "module",
   "scripts": {
-    "build": "remix build",
-    "dev": "remix dev --manual",
+    "build": "tsc && remix build",
+    "dev": "remix dev",
     "lint": "eslint --ignore-path .gitignore --cache --cache-location ./node_modules/.cache/eslint .",
-    "start": "remix-serve ./build/index.js",
+    "start": "node dist/server.js",
     "typecheck": "tsc",
     "build-css": "postcss app/tailwind.css -o dist/output.css",
     "watch-css": "tailwindcss -i ./app/tailwind.css -o ./dist/tailwind.css --minify --watch"
@@ -23,9 +23,10 @@
     "@remix-run/css-bundle": "^2.4.1",
     "@remix-run/node": "^2.4.1",
     "@remix-run/react": "^2.4.1",
-    "@remix-run/serve": "^2.4.1",
+    "@remix-run/serve": "^2.10.3",
     "@tanstack/react-table": "^8.11.3",
     "@tanstack/react-virtual": "^3.8.3",
+    "@types/node": "^20.14.12",
     "api": "^6.1.1",
     "aws-sdk": "^2.1653.0",
     "bcrypt": "^5.1.1",
@@ -54,7 +55,9 @@
     "react-dom": "^18.2.0",
     "tailwind-merge": "^2.2.0",
     "tailwindcss-animate": "^1.0.7",
-    "tiny-invariant": "^1.3.3"
+    "tiny-invariant": "^1.3.3",
+    "ts-node": "^10.9.2",
+    "tsconfig-paths": "^4.2.0"
   },
   "devDependencies": {
     "@remix-run/dev": "^2.4.1",
@@ -81,7 +84,7 @@
     "postcss": "^8.4.33",
     "stream": "^0.0.3",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5.1.6",
+    "typescript": "^5.5.4",
     "util": "^0.12.5"
   },
   "engines": {
