@@ -7,20 +7,17 @@ import { Input } from "@/components/ui/input";
 import { ProfileViewModal } from "@/components/profile-view-modal";
 import { getStudentColumns } from "./columns";
 import { Button } from "@/components/ui/button";
-import { authorize, getData } from "app/googlesheetsserver";
+import { staffStudentDataLayer } from "~/data/initializedatalayer.server";
+
 
 export const loader: LoaderFunction = async ({ request, params }) => {
   const session = await requireUser({ request, params });
-  
-  const auth = await authorize();
-  const sheetId = process.env.GOOGLE_SHEETS_ID;
-  if (!sheetId) {
-    throw new Error("GOOGLE_SHEETS_ID environment variable is not set.");
-  }
-  const { students } = await getData(auth, sheetId, "Students");
-  
+
+  const students = await staffStudentDataLayer.getData();
+
   return { students, session };
 }
+
 
 export default function Students() {
   const { students } = useLoaderData();
